+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Dell EMC OpenManage Ansible Modules allows Data Center and IT administrators to 
 
 With the latest release of Dell EMC OpenManage Ansible Modules, the capabilities have improved with support for OpenManage Enterprise. OpenManage Ansible Modules simplifies and automates provisioning, deployment, and updates of PowerEdge servers and modular infrastructure. It allows system administrators and software developers to introduce the physical infrastructure provisioning into their software provisioning stack, integrate with existing DevOps pipelines and manage their infrastructure using version-controlled playbooks, server configuration profiles, and templates in line with the Infrastructure-as-Code (IaC) principles.
 
+# 2.0.10 (March 27, 2020)
+
+  * The new OME module(ome_application_network_proxy) allows to configure a
+   network proxy
+
+  * The new OME module(ome_application_network_address) allows to configure the
+   DNS and an IPv4 or IPv6 network
+
 # 2.0.9 (March 05, 2020)
 
   * The option group_names in the ome_firmware_baseline and ome_firmware_baseline_compliance_info module is replaced with device_group_names.